💄 style: improve input file upload (#3314)
* 💄 style: improve input height

* 💄 style: improve input height

* fix drag
