--- src/app/(main)/chat/(workspace)/@conversation/features/ChatInput/Desktop/Footer/LocalFiles.tsx
@@ -1,10 +1,0 @@
-import { memo } from 'react';
-
-import { EditableFileList } from '@/features/FileList';
-import { useFileStore } from '@/store/file';
-
-export const LocalFiles = memo(() => {
-  const inputFilesList = useFileStore((s) => s.inputFilesList);
-
-  return <EditableFileList items={inputFilesList} padding={0} />;
-});
