+++ src/app/(main)/chat/(workspace)/@conversation/features/ChatInput/Desktop/Footer/index.tsx
@@ -7,20 +7,17 @@ import { memo, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Center, Flexbox } from 'react-layout-kit';
 
-import DragUpload from '@/components/DragUpload';
 import StopLoadingIcon from '@/components/StopLoading';
 import SaveTopic from '@/features/ChatInput/Topic';
 import { useSendMessage } from '@/features/ChatInput/useSend';
-import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/slices/chat';
 import { useChatStore } from '@/store/chat';
 import { chatSelectors, topicSelectors } from '@/store/chat/selectors';
 import { filesSelectors, useFileStore } from '@/store/file';
 import { useUserStore } from '@/store/user';
-import { modelProviderSelectors, preferenceSelectors } from '@/store/user/selectors';
+import { preferenceSelectors } from '@/store/user/selectors';
 import { isMacOS } from '@/utils/platform';
 
-import { LocalFiles } from './LocalFiles';
+import LocalFiles from '../LocalFiles';
 import SendMore from './SendMore';
 
 const useStyles = createStyles(({ css, prefixCls, token }) => {
@@ -51,10 +48,11 @@ const useStyles = createStyles(({ css, prefixCls, token }) => {
 });
 
 interface FooterProps {
+  expand: boolean;
   setExpand?: (expand: boolean) => void;
 }
 
-const Footer = memo<FooterProps>(({ setExpand }) => {
+const Footer = memo<FooterProps>(({ setExpand, expand }) => {
   const { t } = useTranslation('chat');
 
   const { theme, styles } = useStyles();
@@ -75,28 +73,7 @@ const Footer = memo<FooterProps>(({ setExpand }) => {
 
   const isImageUploading = useFileStore(filesSelectors.isImageUploading);
 
-  const model = useAgentStore(agentSelectors.currentAgentModel);
-
-  const enabledFiles = useUserStore(modelProviderSelectors.isModelEnabledFiles(model));
-
-  const [useCmdEnterToSend, canUpload] = useUserStore((s) => [
-    preferenceSelectors.useCmdEnterToSend(s),
-    modelProviderSelectors.isModelEnabledUpload(model)(s),
-  ]);
-
-  const uploadFile = useFileStore((s) => s.uploadFile);
-
-  const uploadImages = async (fileList: FileList | undefined) => {
-    if (!fileList || fileList.length === 0) return;
-
-    const pools = Array.from(fileList).map(async (file) => {
-      // skip none-file items
-      if (!file.type.startsWith('image') && !enabledFiles) return;
-      await uploadFile(file);
-    });
-
-    await Promise.all(pools);
-  };
+  const [useCmdEnterToSend] = useUserStore((s) => [preferenceSelectors.useCmdEnterToSend(s)]);
 
   const sendMessage = useSendMessage();
 
@@ -142,12 +119,7 @@ const Footer = memo<FooterProps>(({ setExpand }) => {
       padding={'0 24px'}
     >
       <Flexbox align={'center'} gap={8} horizontal style={{ overflow: 'hidden' }}>
-        {canUpload && (
-          <>
-            <DragUpload enabledFiles={enabledFiles} onUploadFiles={uploadImages} />
-            <LocalFiles />
-          </>
-        )}
+        {expand && <LocalFiles />}
       </Flexbox>
       <Flexbox align={'center'} flex={'none'} gap={8} horizontal>
         <Flexbox
