+++ src/app/(main)/chat/(workspace)/@conversation/features/ChatInput/Desktop/index.tsx
@@ -14,6 +14,7 @@ import { systemStatusSelectors } from '@/store/global/selectors';
 
 import Footer from './Footer';
 import Head from './Header';
+import LocalFiles from './LocalFiles';
 import TextArea from './TextArea';
 
 const DesktopChatInput = memo(() => {
@@ -25,33 +26,37 @@ const DesktopChatInput = memo(() => {
   ]);
 
   return (
-    <DraggablePanel
-      fullscreen={expand}
-      headerHeight={HEADER_HEIGHT}
-      maxHeight={CHAT_TEXTAREA_MAX_HEIGHT}
-      minHeight={CHAT_TEXTAREA_HEIGHT}
-      onSizeChange={(_, size) => {
-        if (!size) return;
+    <>
+      {!expand && <LocalFiles padding={'8px 16px'} />}
+      <DraggablePanel
+        fullscreen={expand}
+        headerHeight={HEADER_HEIGHT}
+        maxHeight={CHAT_TEXTAREA_MAX_HEIGHT}
+        minHeight={CHAT_TEXTAREA_HEIGHT}
+        onSizeChange={(_, size) => {
+          if (!size) return;
 
-        updatePreference({
-          inputHeight: typeof size.height === 'string' ? Number.parseInt(size.height) : size.height,
-        });
-      }}
-      placement="bottom"
-      size={{ height: inputHeight, width: '100%' }}
-      style={{ zIndex: 10 }}
-    >
-      <Flexbox
-        gap={8}
-        height={'100%'}
-        padding={'12px 0 16px'}
-        style={{ minHeight: CHAT_TEXTAREA_HEIGHT, position: 'relative' }}
+          updatePreference({
+            inputHeight:
+              typeof size.height === 'string' ? Number.parseInt(size.height) : size.height,
+          });
+        }}
+        placement="bottom"
+        size={{ height: inputHeight, width: '100%' }}
+        style={{ zIndex: 10 }}
       >
-        <Head expand={expand} setExpand={setExpand} />
-        <TextArea setExpand={setExpand} />
-        <Footer setExpand={setExpand} />
-      </Flexbox>
-    </DraggablePanel>
+        <Flexbox
+          gap={8}
+          height={'100%'}
+          padding={'12px 0 16px'}
+          style={{ minHeight: CHAT_TEXTAREA_HEIGHT, position: 'relative' }}
+        >
+          <Head expand={expand} setExpand={setExpand} />
+          <TextArea setExpand={setExpand} />
+          <Footer expand={expand} setExpand={setExpand} />
+        </Flexbox>
+      </DraggablePanel>
+    </>
   );
 });
